requirements.txt
# Core libraries
requests>=2.31.0
spotipy>=2.22.1
beautifulsoup4>=4.12.2
selenium>=4.9.0
rich>=13.7.0
textual>=0.41.0
Pillow>=10.1.0

# Text processing
fuzzywuzzy>=0.18.0
python-levenshtein>=0.23.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
dataclasses-json>=0.6.2
tqdm>=4.66.1
webdriver-manager>=3.8.6

# Development
pytest>=7.4.3
pytest-mock>=3.12.0
black>=23.11.0
isort>=5.12.0 

# Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.1
websockets==12.0 